Add the concat container
This changes the way prometheus works. It now loads a volume that
contains a config file generated from file fragments that are then
combined via the config-concat container into a complete prometheus
file that it can then use to run.

This is a big departure from the original, much simplier, configuration
workflow but it's one that is needed to allow configuring parts of prometheus,
such as alertmanager, from different composible directories.