Require a password for redis access
This PR forces redis to require a password before you can send
it commands and configures the redis exporter to send that
password so it can continue to export the metrics.

The password itself is defined in plain text inside the repo so it
should be considered insecure but it's an improvement over the redis
instance being completely open.